+++ classes/suggested-tasks/local-tasks/providers/content/class-review.php
@@ -5,14 +5,14 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Content;
 
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Local_Task_Factory;
 
 /**
  * Add tasks for content updates.
  */
-class Content_Review extends Content_Abstract {
+class Review extends Content {
 
 	/**
 	 * The provider ID.
@@ -133,6 +133,12 @@ class Content_Review extends Content_Abstract {
 					'post_id' => $post->ID, // @phpstan-ignore-line property.nonObject
 				]
 			);
+
+			// Don't add the task if it was completed.
+			if ( true === \progress_planner()->get_suggested_tasks()->was_task_completed( $task_id ) ) {
+				continue;
+			}
+
 			$items[] = $this->get_task_details( $task_id );
 		}
 		return $items;