+++ classes/suggested-tasks/local-tasks/providers/one-time/class-blog-description.php
@@ -5,12 +5,12 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 /**
  * Add tasks for Core blogdescription.
  */
-class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
+class Blog_Description extends One_Time {
 
 	/**
 	 * The provider ID.
@@ -27,12 +27,19 @@ class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
 	const TYPE = 'configuration';
 
 	/**
-	 * Check if the task condition is met.
+	 * Whether the task is an onboarding task.
+	 *
+	 * @var bool
+	 */
+	protected $is_onboarding_task = true;
+
+	/**
+	 * Check if the task should be added.
 	 *
 	 * @return bool
 	 */
-	public function check_task_condition() {
-		return '' !== \get_bloginfo( 'description' ) ? true : false;
+	public function should_add_task() {
+		return '' === \get_bloginfo( 'description' );
 	}
 
 	/**
@@ -49,18 +56,22 @@ class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
 		}
 
 		return [
-			'task_id'     => $task_id,
-			'title'       => \esc_html__( 'Set tagline', 'progress-planner' ),
-			'parent'      => 0,
-			'priority'    => 'high',
-			'type'        => $this->get_provider_type(),
-			'points'      => 1,
-			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'options-general.php' ) ) : '',
-			'description' => '<p>' . sprintf(
+			'task_id'      => $task_id,
+			'title'        => \esc_html__( 'Set tagline', 'progress-planner' ),
+			'parent'       => 0,
+			'priority'     => 'high',
+			'type'         => $this->get_provider_type(),
+			'points'       => 1,
+			'url'          => $this->capability_required() ? \esc_url( \admin_url( 'options-general.php?pp-focus-el=' . $task_id ) ) : '',
+			'description'  => '<p>' . sprintf(
 				/* translators: %s:<a href="https://prpl.fyi/set-tagline" target="_blank">tagline</a> link */
 				\esc_html__( 'Set the %s to make your website look more professional.', 'progress-planner' ),
 				'<a href="https://prpl.fyi/set-tagline" target="_blank">' . \esc_html__( 'tagline', 'progress-planner' ) . '</a>'
 			) . '</p>',
+			'link_setting' => [
+				'hook'   => 'options-general.php',
+				'iconEl' => 'th:has(+td #tagline-description)',
+			],
 		];
 	}
 }