+++ classes/suggested-tasks/local-tasks/providers/one-time/class-sample-page.php
@@ -5,12 +5,12 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 /**
  * Add tasks to check if WP debug is enabled.
  */
-class Sample_Page extends Local_OneTime_Tasks_Abstract {
+class Sample_Page extends One_Time {
 
 	/**
 	 * The provider type.
@@ -34,6 +34,13 @@ class Sample_Page extends Local_OneTime_Tasks_Abstract {
 	protected $capability = 'edit_pages';
 
 	/**
+	 * Whether the task is an onboarding task.
+	 *
+	 * @var bool
+	 */
+	protected $is_onboarding_task = true;
+
+	/**
 	 * The sample page.
 	 *
 	 * @var \WP_Post|null|false
@@ -41,14 +48,12 @@ class Sample_Page extends Local_OneTime_Tasks_Abstract {
 	protected $sample_page = false;
 
 	/**
-	 * Check if the task condition is met.
+	 * Check if the task should be added.
 	 *
 	 * @return bool
 	 */
-	public function check_task_condition() {
-		$sample_page = $this->get_sample_page();
-
-		return null === $sample_page ? true : false;
+	public function should_add_task() {
+		return null !== $this->get_sample_page();
 	}
 
 	/**